+++ features/support/app_pages/admin/admin_users.rb
@@ -175,4 +175,7 @@ include JavascriptAlerts
     find("input[type='search']").set string
   end
 
+  def delete_wallet_test_user
+    binding.pry
+  end 
 end